+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.2] - add DecoratorRule.words
+
+* add DecoratorRule.words to match multiple words
+
 ## [1.0.1] - add parameters to DecoratorRule.url and DecoratorRule.word
 
 * add looseUrl param to DecoratorRule.url